requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-postgres==5.6.0
psycopg2-binary